fix: toggle visibility of rows when section header is clicked
